Moved Customer.salutation from shop framework into merchant implementation
